requirement.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2026.1
